rename <Parameters> to <SyncedParametersList>
