+++ frontend/src/metabase/query_builder/components/__mocks__/NativeQueryEditor.jsx
@@ -1,10 +1,10 @@
 /* eslint-disable react/prop-types */
 import React from "react";
 
-import Parameters from "metabase/parameters/components/Parameters/Parameters";
+import SyncedParametersList from "metabase/parameters/components/SyncedParametersList/SyncedParametersList";
 
 const MockNativeQueryEditor = ({ location, query, setParameterValue }) => (
-  <Parameters
+  <SyncedParametersList
     parameters={query.question().parameters()}
     query={location.query}
     setParameterValue={setParameterValue}
